Update upgrade flow modals
Based on the Figma views.
