+++ src/components/Modal/UpgradeToTModal/TransactionIdle.tsx
@@ -1,36 +1,40 @@
-import React, { FC } from "react"
+import { FC } from "react"
 import {
-  Alert,
-  AlertIcon,
-  AlertTitle,
   Box,
+  Button,
   HStack,
   ModalBody,
   ModalCloseButton,
+  ModalFooter,
   ModalHeader,
 } from "@chakra-ui/react"
-import TransactionError from "../../../static/icons/TransactionError"
-import ModalFooter from "./ModalFooter"
-import { Body1, H5 } from "../../Typography"
+import { Body1, Body3, H5 } from "../../Typography"
 import UpgradeIconGroup from "../../UpgradeIconGroup"
 import { Token } from "../../../enums"
 import TransactionStats from "./TransactionStats"
 import { Divider } from "../../Divider"
+import ViewInBlockExplorer from "../../ViewInBlockExplorer"
+import { ExplorerDataType } from "../../../utils/createEtherscanLink"
+import { BaseModalProps } from "../../../types"
+import withBaseModal from "../withBaseModal"
 
-interface TransactionIdleProps {
+interface TransactionIdleProps extends BaseModalProps {
   upgradedAmount: number
   receivedAmount: number
   exchangeRate: number
-  transactionId: string
   token: Token
+  vendingMachineContractAddress: string
+  onSubmit: () => void
 }
 
 const TransactionIdle: FC<TransactionIdleProps> = ({
   upgradedAmount,
   receivedAmount,
   exchangeRate,
-  transactionId,
   token,
+  vendingMachineContractAddress,
+  onSubmit,
+  closeModal,
 }) => {
   return (
     <>
@@ -54,10 +58,21 @@ const TransactionIdle: FC<TransactionIdleProps> = ({
           receivedAmount={receivedAmount}
           upgradedAmount={upgradedAmount}
         />
-        <ModalFooter transactionId={transactionId} />
+        <Body3>
+          This action i reversible via the{" "}
+          <ViewInBlockExplorer
+            id={vendingMachineContractAddress}
+            type={ExplorerDataType.ADDRESS}
+            text="vending machine contract."
+          />
+        </Body3>
       </ModalBody>
+      <ModalFooter>
+        <Button onClick={closeModal}>cancel</Button>
+        <Button onClick={onSubmit}>upgrade</Button>
+      </ModalFooter>
     </>
   )
 }
 
-export default TransactionIdle
+export default withBaseModal(TransactionIdle)
